unicore/Name.pl changed in 5.14 and breaks meta unicode List
Thu May 26 12:24:03 2011: Request 68488 was acted upon.
Transaction: Ticket created by RWSTAUNER
       Queue: Acme-MetaSyntactic
     Subject: unicore/Name.pl changed in 5.14 and breaks meta unicode List
   Broken in: 0.99
    Severity: Normal
       Owner: Nobody
  Requestors: RWSTAUNER@cpan.org
      Status: new
 Ticket <URL: https://rt.cpan.org/Ticket/Display.html?id=68488 >

t/17import.t breaks on perl 5.14 when testing the "unicode" module.

The unicore/Name.pl file has changed format
(at least since 5.10 which was my previous perl version).

There is now only 1 tab where there used to be two, I think.
It looks to me like the one in 5.14 doesn't have any of the lines that
were being skipped in the older versions.

I have attached a patch with alternate logic for skipping "blocks"
so that the unicode list ends up with 23085 instead of 0 entries.

I tested it with the Name.pl from 5.14, 5.10.1, and 5.6.1
and it appeared to work for me with all of them.