the extract() method for MultiList themes with a remote list now accept a second parameter
This second parameter is the theme category, which now allows
the extract() method to process different source URL differently.

Caveat: the code does a reverse mapping of URL to category,
so it will break when two categories have the same URL but
process it differentely (a very likely scenario).

This needs to be fixed.