Deduplicate the list passed to init()
The deduplication is already done in Acme::MetaSyntactic::MultiList,
because intermediate nodes on the hierachy tree are automatically
created and filled by merging the lists of names from deeper levels.

Deduplicating the simple lists makes Acme::MetaSyntactic more robust
and more consistent. This also makes the subtest_uniq() test in
Test::MetaSyntatic irrelevant.

Thanks to Abigail for noting this.