prevent CPAN.pm from detecting the recursive loading during installation
In some versions, CPAN.pm prints a spurious "\n" in the code that
handles the "CPAN.pm loaded withing CPAN.pm" recursion (using the
PERL5_CPAN_IS_RUNNING environment variable).

This happens when the pause_id theme is loaded as part of:

    use Acme::MetaSyntactic ':all';

which makes the t/50usecase.t test fail because of the extra "\n"
in the output.

By deleting the PERL5_CPAN_IS_RUNNING in the use case, we ensure that
CPAN.pm recursion detection is not triggered, and fix the script output.